# apps/backend/rust-api/openapi.yaml
openapi: 3.0.3
info:
  title: BeamLab Advanced Analysis API
  version: 2.1.0
  description: High-performance structural analysis endpoints for modal, time-history, and seismic analyses.
servers:
  - url: http://localhost:8000
paths:
  /api/analysis/modal:
    post:
      summary: Modal eigenvalue analysis
      operationId: modalAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModalAnalysisRequest'
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModalAnalysisResponse'
        '400':
          description: Invalid input
        '500':
          description: Analysis failed
  /api/analysis/time-history:
    post:
      summary: Time-history dynamic analysis
      operationId: timeHistoryAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimeHistoryRequest'
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeHistoryResponse'
        '400':
          description: Invalid input
        '500':
          description: Analysis failed
  /api/analysis/seismic:
    post:
      summary: Seismic response spectrum analysis
      operationId: seismicAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeismicAnalysisRequest'
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeismicAnalysisResponse'
        '400':
          description: Invalid input
        '500':
          description: Analysis failed
components:
  schemas:
    ModalAnalysisRequest:
      type: object
      properties:
        stiffness_matrix:
          type: array
          items:
            type: number
        mass_matrix:
          type: array
          items:
            type: number
        dimension:
          type: integer
        num_modes:
          type: integer
          default: 10
        mass_type:
          type: string
          enum: [Consistent, Lumped]
          default: Consistent
        normalize_modes:
          type: boolean
          default: true
        compute_participation:
          type: boolean
          default: true
      required: [stiffness_matrix, mass_matrix, dimension]
    ModalAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        frequencies_hz:
          type: array
          items:
            type: number
        frequencies_rad_s:
          type: array
          items:
            type: number
        periods_s:
          type: array
          items:
            type: number
        mode_shapes:
          type: array
          items:
            type: array
            items:
              type: number
        modal_masses:
          type: array
          items:
            type: number
        participation_factors:
          type: array
          items:
            type: number
        cumulative_participation:
          type: array
          items:
            type: number
        performance_ms:
          type: number
    TimeHistoryRequest:
      type: object
      properties:
        stiffness_matrix:
          type: array
          items:
            type: number
        mass_matrix:
          type: array
          items:
            type: number
        dimension:
          type: integer
        force_history:
          type: array
          items:
            type: array
            items:
              type: number
        dt:
          type: number
        initial_displacement:
          type: array
          items:
            type: number
          nullable: true
        initial_velocity:
          type: array
          items:
            type: number
          nullable: true
        integration_method:
          type: string
          enum: [newmark, central_difference, wilson]
          default: newmark
        damping:
          type: object
          properties:
            type:
              type: string
              enum: [none, rayleigh, modal]
              default: none
            alpha:
              type: number
              nullable: true
            beta:
              type: number
              nullable: true
            ratios:
              type: array
              items:
                type: number
              nullable: true
          required: [type]
        output_interval:
          type: integer
          default: 1
      required: [stiffness_matrix, mass_matrix, dimension, force_history, dt]
    TimeHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        time:
          type: array
          items:
            type: number
        displacement_history:
          type: array
          items:
            type: array
            items:
              type: number
        velocity_history:
          type: array
          items:
            type: array
            items:
              type: number
        acceleration_history:
          type: array
          items:
            type: array
            items:
              type: number
        max_displacement:
          type: number
        max_velocity:
          type: number
        max_acceleration:
          type: number
        performance_ms:
          type: number
    SeismicAnalysisRequest:
      type: object
      properties:
        frequencies_rad_s:
          type: array
          items:
            type: number
        mode_shapes:
          type: array
          items:
            type: array
            items:
              type: number
        modal_masses:
          type: array
          items:
            type: number
        participation_factors:
          type: array
          items:
            type: number
        seismic_code:
          type: string
          enum: [IS1893, ASCE7, EC8]
        zone:
          type: string
        soil_type:
          type: string
        importance:
          type: string
        response_reduction:
          type: string
        damping_ratio:
          type: number
          default: 0.05
        combination_method:
          type: string
          enum: [SRSS, CQC, ABS]
          default: CQC
        story_heights:
          type: array
          items:
            type: number
          nullable: true
        story_masses:
          type: array
          items:
            type: number
          nullable: true
      required: [frequencies_rad_s, mode_shapes, modal_masses, participation_factors, seismic_code, zone, soil_type, importance, response_reduction]
    SeismicAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        periods_s:
          type: array
          items:
            type: number
        spectral_accelerations_g:
          type: array
          items:
            type: number
        modal_displacements_m:
          type: array
          items:
            type: number
        modal_base_shears_kn:
          type: array
          items:
            type: number
        max_displacement_m:
          type: number
        max_base_shear_kn:
          type: number
        code_base_shear_kn:
          type: number
        story_forces:
          type: array
          items:
            type: object
            properties:
              level:
                type: integer
              height_m:
                type: number
              lateral_force_kn:
                type: number
              cumulative_shear_kn:
                type: number
        combination_method:
          type: string
        performance_ms:
          type: number